+++ eclipsePlugin/src/de/tobject/findbugs/properties/FindbugsPropertyPage.java
@@ -25,7 +25,7 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
-import javax.annotation.Nonnull;
+import jakarta.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 import org.eclipse.core.resources.IProject;