+++ spotbugs/src/main/java/edu/umd/cs/findbugs/BugRankCategory.java
@@ -19,7 +19,7 @@
 
 package edu.umd.cs.findbugs;
 
-import javax.annotation.Nonnull;
+import jakarta.annotation.Nonnull;
 
 /**
  * Smaller value is scarier