+++ spotbugs/src/main/java/edu/umd/cs/findbugs/ba/CFGBuilderFactory.java
@@ -19,7 +19,7 @@
 
 package edu.umd.cs.findbugs.ba;
 
-import javax.annotation.Nonnull;
+import jakarta.annotation.Nonnull;
 
 import org.apache.bcel.generic.MethodGen;