+++ spotbugs/src/main/java/edu/umd/cs/findbugs/detect/NoiseNullDeref.java
@@ -27,7 +27,7 @@ import java.util.Set;
 import java.util.SortedSet;
 
 import javax.annotation.CheckForNull;
-import javax.annotation.Nonnull;
+import jakarta.annotation.Nonnull;
 
 import org.apache.bcel.classfile.JavaClass;
 import org.apache.bcel.classfile.LineNumberTable;