+++ spotbugs/src/main/java/edu/umd/cs/findbugs/util/ClassPathUtil.java
@@ -24,7 +24,7 @@ import java.util.StringTokenizer;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import javax.annotation.Nonnull;
+import jakarta.annotation.Nonnull;
 
 /**
  * Some utility methods for working with the Java class path.