+++ spotbugsTestCases/src/java/nullnessAnnotations/UseAnnotations.java
@@ -1,7 +1,7 @@
 package nullnessAnnotations;
 
 import javax.annotation.CheckForNull;
-import javax.annotation.Nonnull;
+import jakarta.annotation.Nonnull;
 
 public class UseAnnotations {