+++ spotbugsTestCases/src/java/sfBugs/Bug3424096.java
@@ -1,6 +1,6 @@
 package sfBugs;
 
-import javax.annotation.Nonnull;
+import jakarta.annotation.Nonnull;
 
 
 public class Bug3424096 {